+++ NEWS.md
@@ -1,5 +1,9 @@
 # LAGOSUS 0.0.2
 
+## MINOR IMPROVEMENTS
+
+ * `coordinatize` function has a default projection (#12)
+
 # LAGOSUS 0.0.1
 
 * Initial prototype release